WIP: core/cortex-m/ec.lds.S: Fix .data placement when no flash is present
When CONFIG_FLASH_PHYSICAL is not set (e.g. on kukui_scp that only
has RAM), .bss and the initial values for .data end up overlapping.

This fixes the issues by placing that region after .bss/.bss.slow
in that case.

TODO: This means that reboot is broken, as the EC only has a
single copy of .data, and is not able to restore to the initial
values on restart.

BRANCH=none
BUG=b:122084384
TEST=Flash kukui_scp, no more strange crashes on boot

Change-Id: Ifaed35f18923f1a73bb7537288d11ff1eeee5bab
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>