cr50: Deferred RDD check in deasserting EC_RST_L.
RDD connection can be detected after the debug cables comparator
values are stable, which takes RDD_MAX_WAIT_TIME_COUNTER.

This CL changes to check RDD connection in a deferred way only when
it is a power-on reset and the power button is held.

This CL increases Flash memory usage by 72 bytes.

BRANCH=cr50
BUG=b:37351386
TEST=manually on bob, coral and scarlet.

1. (EC) cutoff
2. ubplug all cables.
3. hold the power button.
4. plug USB-C cable of Servo V4 (or SuzyQ cable) in.
5. check EC console. Should be off.
   To verify, run the cr50 console command, 'ecrst' and check
   if 'EC_RST_L' is asserted.
6. release the power button. EC should run.

Change-Id: I4c7b0653bf49fbe57c8ad8395c72e006fa3a0ff0
Signed-off-by: Namyoon Woo <namyoon@chromium.org>