oz554: Keep interrupt always enabled
OZ554ALN asserts the interrupt when it's ready for writing settings
and settings are cleared when it's turned off (i.e. S3/S5).

Therefore, we do not need to toggle the interrupt on power state
transitions. This patch makes the interrupt enabled on HOOK_INIT.

It's assumed the device doesn't have a lid. For clamshell devices,
different interrupt & power control scheme may be needed.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b/120237453
BRANCH=Kalista
TEST=Verified display backlight is turned on/off as system changes
power states.

Change-Id: Ia911993abd76c09e79025ea454e456781b5b2fc1