octopus: convert boards to use hardware int
Recent kernel changes expect the EC to use a dedicated interrupt pin
from the EC to the AP to notify the AP of pending sensor data (instead
of using an eSPI "interrupt").

The octopus boards have this hardware support, we just need to enable
the EC use it.

BRANCH=octopus
BUG=b:122552125
TEST=perform sensor tests on various octopus boards

Change-Id: I2bd3ffe14947d5f1ec71acbb53fcac962b007cf9
Signed-off-by: Jett Rink <jettrink@chromium.org>