nocturne: Check DP MF-bit against selected pin cfg
When we are configuring a Type-C port for DisplayPort alternate mode, we
should check to see that the selected pin config supports multi-function
mode or not.  This commit fixes a bug where we were setting the
SuperSpeed muxes based solely upon the Multi-function Preferred bit in
the DPStatus VDO.  Some Type-C video adapters are buggy and set the MF
preferred bit without actually supporting an MF pin configuration.
Therefore, we trust the reported supported pin configurations in the
DiscMode VDO.

BUG=chromium:919756
BRANCH=firmware-nocturne-10984.B
TEST=Flash nocturne, plug in Insignia NS-PU369CH-WH USB-C to HDMI
adapter, verify that 4k60 display is shown.
TEST=Plug in Belkin dock which supports SuperSpeed ports, verify that
SuperSpeed ports work and display is shown at 4k30.

Change-Id: I80112d103376c4173691d85a61733cbe154e8730
Signed-off-by: Aseda Aboagye <aaboagye@google.com>