cheza: Ensure POWER_GOOD low after forcing shutdown
In power_chipset_init(), it forces the system shutdown by turning off the
switchcap if it is not a sysjump. The POWER_GOOD indicator needs time to
drop. Make sure it is low before executing the power state machine.

Make this wait to every case of turning off the swithcap.

BRANCH=none
BUG=b:120889835
TEST=Typed "reboot ap-off" in EC console and verified AP off.

Change-Id: I50f4f51793de1482fc2dd9b3dd05819bb2501cdc
Signed-off-by: Wai-Hong Tam <waihong@google.com>