hatch: Add support for MST (multi stream transport) enable
The MST chip for Hatch needs to be enabled when HPD signal from either
the Port 1 TCPC or HDMI port is high. This CL adds support to to
enable the MST chip based on this criteria. For the Port 1 type C
port, the HPD signal level is derived from the USB PD policy level
where the HPD update driver method is called.

BRANCH=none
BUG=b:123894908
TEST=Used external HP Z27n monitor and verifed the display is extended
as expected when it's connected to either port 1 type C port or the
HDMI connector.

Change-Id: I1c46534bc7f32221f9e379dd9c74d5618c8f57e1
Signed-off-by: Scott Collyer <scollyer@google.com>