cometlake: Guard passing RSMRST_L to PCH by PP5000_A rail
RSMRST_L gets passed along to the AP by EC when it changes state. For
high to low transitions the EC needs to ensure that the PP5000_A rail
is up prior to passing the RSMRST_L transition to the AP.

This CL adds a check to prevent calling
common_intel_x86_handle_rsmrst() when RSMRST_L is high if the PP5000_A
rail is not up. Since PP5000_PG signal will float high when the
regulator is not powered at all, both the enable and power good
signals are used to verify that PP5000_A rail is powered.

BRANCH=none
BUG=b:122631914
TEST=Verified on scope that RSMRST_L to PCH rising edge happens after
PP5000_A rail rising.

Change-Id: Icfd4dabbdfaf6ae76b3cdcbc6f75a5188a21ff51
Signed-off-by: Scott Collyer <scollyer@google.com>