Update patch set 2
Patch Set 2: Code-Review-1

If the AP is in S0, you need to send the keyboard event to chrome, even in tablet mode.
There is a mode "magic keyboard" when chrome needs the events if an external screen is connected and a the main screen brightness is set to 0.

Patch-set: 2
Reviewer: Gerrit User 1104704 <1104704@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1