cr50: Enable LTO
Tpm2 library still can not be compiled with LTO enabled, as it results
in untraceable .bss segment, which in turn makes it impossible to
separate .bss segment of the Tpm2 library into a contiguous memory
section.

As presented, enabling LTO on Cr50 saves 4372 bytes of flash space.

BRANCH=cr50, cr50-mp
BUG=b:65253310
TEST=after the entire stack was applied Cr50 builds successfully with
     CONFIG_LTO enabled, and passes TCG test suite.

Change-Id: I4ea02634898166be765bb0986186eb4ecdadd511
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>