USB-PD: Disable Vconn on SRC_DISCONNECT
Currently, source only devices do not turn off Vconn when a partner
is disconnected. This causes the polarity detection to be disabled.

This patch turns off Vconn on SRC_DISCONNECT for
!CONFIG_USB_PD_DUAL_ROLE boards.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b/122861749
BRANCH=Kalista
TEST=Plug in an ethernet dongle via a USB-C-to-A adapter. Replug it
upside down and verify Ethernet is turned on on both sides.

Change-Id: I4ac303f2121d9e4114ce7097db304e4f65f27ae0
Reviewed-on: https://chromium-review.googlesource.com/1412228
Commit-Ready: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Zhuohao Lee <zhuohao@chromium.org>