LSM6: Correct timestamp reporting and FIFO lengths
When an accelerometer sends a sample to the motion_sense task with a
timestamp, the timestamp should represent the time of the IRQ which
caused the FIFO read.  This change caches a timestamp after the
accelerometer data is read to use when pushing all data from that read.

This also corrects the number of bits in the FIFO length field and the
units for the FIFO watermark.

BRANCH=octopus
BUG=b:120679547
TEST=builds, phaser360 CTS pass rate unaffected

Change-Id: I220aa2e8fa23af3f7833999cdfac966e8695c831
Signed-off-by: Diana Z <dzigterman@chromium.org>