WIP it83xx/spi: add spi slave function
Add the spi slave function which is required to
communicate with the EC when the CPU is the ARM
processor.

This spi slave function is used to receive data
by DMA mode. When the data transfer is completed,
the interrupt signal will generate, and the data
will pass to common layer to handle request.

BUG=none
BRANCH=none
TEST=Connect the master to the slave and transfer the
     test command which is defined in host_command.c.
     And then corresponding response message is obtained
     from the console.

Change-Id: I4fcb952c47ad7c55d2d0ccbbaa145baaeb129bdb
Signed-off-by: tim <tim2.lin@ite.corp-partner.google.com>