nocturne: Check IRQ and HPD in DPStatus message.
Per the VESA DisplayPort Alt Mode on USB Type-C spec, IRQ_HPD indicates
that a high to low followed by a low to high transition was detected.
Therefore, we should be checking when IRQ is high and HPD is low is
received as that is an error.  This commit fixes that bug where were
comparing our level to the GPU instead of what was shown in the PDO.

BUG=none
BRANCH=firmware-nocturne-10984.B
TEST=Flash nocturne, plug in Apple Multiport AV USB-C adapter, verify
that external display works.

Change-Id: I536dd0d2556794411eb8d54a9a0c34f61fff5253
Signed-off-by: Aseda Aboagye <aaboagye@google.com>