rammus: turn the USB type-a charging off by default in suspend mode
This patch turns the USB type-a charging off by default in suspend mode.
If the DUT power on with connecting a USB device, the OS must send an
event to clear the 8th-bit. Otherwise, the USB device will lost the
power in suspend mode, and the wakeup event won't be generated if the
device is a wakeup source (eg. USB keyboard or mouse).

BUG=b:121438672
BRANCH=firmware-rammus-11275
CQ-DEPEND=CL:1423758
TEST=Power on DUT without connecting USB, then suspend the DUT.
     Verfied the USB type-a charging off.

Change-Id: I39a0f85d7033b798a80e80cc302a2c72cebae8d6
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>