npcx7: clock: increase core clock to 90MHz during EC booting
It was observed that the security verification is the bottleneck for the
EC boot time. In this CL, we reduce this bottleneck by introducing a new
configuration flag "CONFIG_CLOCK_TURBO_AT_BOOTUP." If the flag is defined,
the core clock frequency will increase to 90MHz during EC booting until
the vboot hash finishes. The method to check if the vboot hash is in
progress or not follows the CL:326792.
This solution applicable to NPCX7 and later chips.

BRANCH=none
BUG=none
TEST=
-- For normal VBOOT_HASH --
1. test on npcx7_evb with 192 KB RW image:
               |  current  | with this CL|
---------------+-----------+--------------
Hash time (ms) |   1264.5  |    311.9    |

2. test on npcx7_evb with 320 KB RW image:
               |  current  | with this CL|
---------------+-----------+--------------
hash time (ms) |   2104.7  |    519.7    |

3. test on a real platform (YORP Proto 2):
a. currently, hash done at timestamp -> 1.201643 S
  [1.201643 hash done 5a9bed44b50ad770e7ed608ac065c013d0d09171bd0d469cac0cd4091402f86f]
b. apply this CL, hash done at timestamp -> 0.502152
  [0.502152 hash done 9bc5f8a1a40b80621af03c3e8bae6af0d7e7e81ab673d43f73dbf3fd2fce4014]

-- For EFS (HASH and RSA3072) --
1. test on npcx7_evb with FIZZ image (128260 bytes)
                 |  current  | with this CL|
-----------------+-----------+--------------
Hash time (ms)   |   651.1   |    158.8    |
-----------------+-----------+--------------
RSA3072 time (ms)|   228.1   |     58.8    |
-----------------+-----------+--------------
Total time (ms)  |   879.1   |    217.6    |

Change-Id: Ibb2a16e97ae2216d4921766b917bed420116fc26
Signed-off-by: CHLin <CHLIN56@nuvoton.com>