npcx7: clock: increase core clock to 90MHz during EC booting
It was observed that the security verification is the bottleneck for the
EC boot time. In this CL, we reduce this bottleneck by introducing a new
configuration flag "CONFIG_CLOCK_TURBO_AT_BOOTUP." If the flag is
defined, the core clock frequency will increase to 90MHz during EC
booting until the vboot hash finishes. The method to check whether the
vboot hash is in progress follows the CL:326792.
This solution is applicable to NPCX7 and later chips.

BRANCH=none
BUG=none
TEST=No error for "make buildall"
TEST=test on npcx7_evb with this CL, we got ~4x improvement of
security verification time.

Change-Id: Ibb2a16e97ae2216d4921766b917bed420116fc26
Signed-off-by: CHLin <CHLIN56@nuvoton.com>