tcpm/it83xx.c: the USB-PD PHY trimming value is reloaded wrongly after software reset
  For every chip, the USB-PD PHY characteristic is trimmed to
meet the USB-IF spec. These parameters are stored in a non-volatile
region and load to DFF when EC power on or hardware reset. But to
load the parameter, clk_fnd can't operate at 48Mhz, or this will
cause USB-PD TX eye signal getting worse on Dx after triggering
software reset. So we don't reload cc parameter settings during
initialization for Dx.

BUG=b:123173731
BRANCH=none
TEST=cc parameters are the same, after ec software reset and
hardware reset on Bx and Dx.

Change-Id: I0c7144afbb648680bb60cc128c8212abb43ccd68
Signed-off-by: Ruibin Chang <Ruibin.Chang@ite.com.tw>