ish: fix timer 0 periodic mode
We are using timer 0 with periodic mode and
in this mode, the value of timer 0 comparator
should be automatically increased by the value
last written when main counter equals the value.

But due to mis-configuration, currently it is not
automatically increased.

This patch corrects the configuration for timer 0.

BRANCH=none
BUG=b:112750896
TEST=Tested on Atlas and verified increased value in
     timer 0 comparator.

Change-Id: I1c98d980cd5c28226e223ca058b2181cf40b87f9