octopus/ampton: disable espi pad when system goes into G3
On Ampton, the EC's power consumption is about 5.5mA on battery
mode in G3 state. This is because EC unable to go into low power
mode properly in idle task when eSPI CS# pin is low.
So we disable eSPI pad when system goes into G3 state to reduce
power number.

BUG=b:121105042
BRANCH=none
TEST=On Ampton, EC power rail drops to about 1.1mA in G3.

Signed-off-by: Dino Li <Dino.Li@ite.com.tw>
Change-Id: I1c85bd0e909936e7f143b15b5e9b7c1884d5cc62