Revert "casta : add battery information"
This reverts commit 3f21cfa6e7800de7a9a0a3247fecd84e0114655c.

Reason for revert: The change needs to be uploaded first to master. Also it has lots of style errors that needs some manual cleanup. Reverting this change to do the cleanup, upload it to master, get it merged in that branch and then cherry-pick it back to this branch

Original change's description:
> casta : add battery information
> 
> casta uses same battery with nautilus.
> 
> BUG=b:122868858
> BRANCH=None
> TEST=flash EC & check battery information
> 
> Change-Id: I55894821744d242958c2dcf31da355b315d9ac8f
> Signed-off-by: YongBeum Ha <ybha@samsung.com>
> Reviewed-on: https://chromium-review.googlesource.com/c/1385764
> Reviewed-by: Philip Chen <philipchen@chromium.org>

Bug: b:122868858
Change-Id: Ia80091841837084093bcc342eb4583a7e2b4820b