cr50: REFRESH+PWR recovery mode state lost to deep sleep
The recovery mode latched state gets lost due to deep sleep if the
the Cr50 console is idle prior to capturing the REFRESH+PWR key combo.
This change delays sleep so the AP has time to power back on and read
the recovery state.

BUG=b:123307572
BRANCH=cr50
TEST=make buildall.  Verified boot to recovery mode screen with servo
disconnected and verified boot to recovery mode with non dev image.
Change-Id: Idf744c6967dc83c26a60429134ce3e8a9ab62ea3
Signed-off-by: Keith Short <keithshort@chromium.org>