Update patch set 13
Patch Set 13: Verified+1

It seems that this just adds too much code for glados and granite, so I've surrounded the code with an ifdef used when enabling a resettable task (see next CL).

Also, I didn't quite get to the bottom of it but the fizz pre-cq failure seems that it was just because the expression was too complicated; I've replaced it with a simpler one which is equivalent and it is happy about.

Patch-set: 13
Reviewer: Gerrit User 1256783 <1256783@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1