Update patch set 1
Patch Set 1:

We will needed this, and then we need to propagate this the ec_command.h header file to the kernel.

I can submit this CL (or you can; either way). If you want to do it, put it on the bottom of your CL stack. (Let me know either way!)

Patch-set: 1