rammus/shyvana: lower the maximum allowed ODR to 100Hz
Some tests are failing on Shyvana at the highest ODR.
This is likely due to Shyvana not having a dedicated GPIO path
for MKBP events, instead relying on the ACPI path which is known
to have lower latency guarantees than we expect.

Lowering the ODR allows the tests to pass more reliably.

BUG=b:123477920
TEST=run failing CTS test cases on Shyvana

Signed-off-by: Enrico Granata <egranata@chromium.org>

Change-Id: I864b90f26017c382e48adf976ecf5d6500480037