ish: move setting start_called to proper place
Some functions(like enabling interrupt) should start after
scheduler is ready and they use start_called which indicates
if scheduler is ready or not.

Currently start_called is set by the first task but it is away
after the task did many things.

So this patch moves the setting start_called to the place where
the scheduler is ready.

BRANCH=none
BUG=none
TEST=verified in Atlas platform

Change-Id: I24d9cec411e91b7365f46fa8daf4a02fe43287dd