make: improve buildall problems reporting
When buildall fails for a reason other than failure to build a board
image, it is difficult to figure out what exactly went wrong, the
amount of output on the screen is overwhelming, the vast majority of
it having nothing to do with the failure.

Also, experience shows that the user does not really care about the
exact build phase when a failure happened. As soon as the user
detects file(s) in .failedboards directory, he/she can try building
the board separately and see what the problem is.

With this patch not only building the boards, but also running regular
tests, fuzz and cts tests will create appropriate files in the
./.failedboards directory when starting the appropriate make step and
will remove the files when succeeding.

Management of ./.failedboards directory is also being simplified: its
creation is enforced before building various targets, but cleaning
happens only when the target is buildall.

Now, after running 'make buildall' files in ./.failedboards give a
better indication of what went wrong:
  - regular files - building board with this name failed
  - files host-XXXX_fuzz - fuzzer XXX_fuzz failed
  - files host-XXXX - test XXXX failed
  - files cts-XXXX-YYY cts test for board XXX test YYY failed

BRANCH=none
BUG=none
TEST=created various failures and observed the appropriate leftovers
     in ./.failedboards

Change-Id: Iadfe1c63da5db7fe37b8647b277b61a461de5399
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1444795
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
(cherry picked from commit 9bcaf16c7de255e2e906bef795f0b62b0a19c124)