include/link_defs: Properly constify pointers
In C, const can be in 2 places: "const void *ptr" if the _data_
pointed by the pointer is constant. "void *const ptr" if the
pointer itself is a constant (but not the data).

Fix entries in link_defs.h to match this.

In passing, this triggered an error in common/tpm_registers.c, where
the memset statement appears to be incorrect.

BRANCH=cr50
BUG=none
TEST=make buildall -j

Change-Id: Icfa714b04ad89f251503af1b44fa2f0c8747d8dc
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>