Update patch set 2
Patch Set 2:

I just created CL:1446539 so we could submit this on ToT before I say that you rebased it. There is one change in it - to base the feature flag on CHIP_ISH instead of HECI. Otherwise they are the same.

I will let you give a +1 to my CL or (update this CL with CHIP_ISH and submit this). Either once it fine with me.

Patch-set: 2