temp_sensor: Guard thermistor readings by their power rail enable gpio
If a temperature reading is taken using a thermistor circuit when the
power rail for that circuit is not powered, then the resultiing
reading will convert to a very high temperature which will then lead
to a false thermal shutdown event.

This CL utilizes an existing config option
CONFIG_TEMP_SENSOR_POWER_GPIO to allow for the thermistor driver
functions to know which gpio to check to see if a power rail is
enabled or not. If the rail is not enabled, the function returns
EC_ERROR_UNKNOWN so the reading will be ignored by the calling
function. Since this change to the thermistor driver is dependent on
the config option, it does not affect in existing boards, but can be
used for boards that have this requirement.

BRANCH=none
BUG=b:123900860
TEST=Verified that the the EC console message 'thermal SHUTDOWN' no
longer appears following an EC reboot as the ADC readings are now
ignored if the GPIO controlling the power rail is not enabled.

Change-Id: I92b010ab7e801f4897edb3a6d0325d9d7fc74f21
Signed-off-by: Scott Collyer <scollyer@google.com>