Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> We should also create a partner issue for this bug and what the failure mode that is caused by not having this CL
> 
I'll file a bug. Not requesting VNN before an access may return bad data or may cause a hang. 

> 
> Also can you explain VNN a little more in that newly created issue?

VNN literally refers to a rail that powers the uncore logic.  In this context, a VNN request refers to a request to Power Management Controller (PMC) in SoC, to power up the path to DDR (fabrics,system agent, etc.). 

ISH can stay on in S0ix state when major blocks in the SoC are clock gated or power gated, and DDR is in self-refresh. In such a state for e.g. if ISH needs to access DDR, it needs to first request the PMC/SoC to power up the path to DDR.

Patch-set: 2
Reviewer: Gerrit User 1146375 <1146375@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>