motion_sense: send ODR events to the AP
This commits enables the EC to send a special event type to the AP
when a sensor's ODR is changed.

This is done in a way that is backwards compatible on the AP side,
in order to enable previous kernels to cooperate (albeit at a loss
of new functionality) with up-to-date ECs.

The plan is to use this information on the kernel side to force
a flush of the sensor ring's state, and avoid potential issues where
the sensor ring uses outdaded information to determine sensor timestamps.

BUG=b:123700100
TEST=run CTS sensor tests, observe the sensor ring not spreading too
far into the future due to a period change

Change-Id: I5a881e9fc5ddb1dbe6690c455d43dc20bffe889a
Signed-off-by: Enrico Granata <egranata@chromium.org>