getversion.sh: use better timestamp algorithm
When running ./util/getversion.sh in a 'dirty' tree do not generate
the timestamp containing the current time, find out outstanding files
(including those not yet included in git) and use the timestamp of the
most recent of these files.

The generated date is not accurate in some cases: files not included
in the tree possibly are not part of the build either, so their
modification should not be causing the generated timestamp change.
Also, if a previously modified file was checked out from git HEAD, it
will not be included in the timestamp string generation.

These inaccuracies are not that important: worst case one of the
builds would have incorrect timestamp, not a big deal for the case
when there are local tree modifications.

The big advantage of the new timestamp generation scheme is that if
there has been no changes in the tree, the new ec_version.h file is
not generated, which prevents re-compilation of board.c. This is
necessary to achieve no rebuilds when make is invoked on a tree which
has not been modified since previous make invocation.

BRANCH=none
BUG=none
TEST=running 'make BOARD=cr50' does not trigger re-compiling of
     board.c any more, the timestamp included in the build is the
     timestamp of the most recent modification of the checked out
     files in the tree

Change-Id: I7c931e4965811faa8079d6762b9b9a57a6bb20bb
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>