Update patch set 1
Patch Set 1:

(7 comments)

> Patch Set 1: Verified+1
> 
> This patch makes sure FPU context saving&restoring only happens for the tasks uses(actually declare to use) FPU H/W
> 
> Tasks which use FPU H/W should have set least significant bit of stack size.
> 
> Currently I'm using the least significant bit of stack size to give hint to task scheduler but I prefer to introduce one more argument like TASK_ALWAYS(n, r, d, s, f).
> 
> I'm not sure how the ECOS grows but I believe it's better to have a mechanism to give more detailed task information which can be more hints for low level components(e.g. scheduler) to do their job efficiently.

Let's use the existing task parameter to signal that we are using FPU or not. In our case (ISH), I would be ok with throwing an expectation if an fpu instruction is encountered on a task that does not have the bit; thoughts?

Patch-set: 1
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>