flash_ec: add "--read" flags for it83xx EC.
"--read" takes a following string as a path to store the image file.
If it is specified, flash_ec reads only an EC firmware image and store
it to the given path.

iteflash verifies the programed EC binary as default, which is
identical to "--verify" design in flash_ec. There is no need to
implement the code handling "--verify" for ITE EC.

CQ-DEPEND=CL:1364327, CL:1481829
BUG=b:123881479
BRANCH=none
TEST=manually ran flash_ec on bip.

[servo_v2 and suzy-Q respectively, on bip]
./util/flash_ec --board bip --read ${IMG_READ}
./util/flash_ec --board bip --image ${IMG} --verify

[servo_v2 and suzy-Q respectively, on dragonegg]
./util/flash_ec --board dragonegg --read ${IMG_READ}
./util/flash_ec --board dragonegg --image ${IMG} --verify

[suzy-Q on scarlet]
./util/flash_ec --board scarlet --read ${IMG_READ}
./util/flash_ec --board scarlet --image ${IMG} --verify

Change-Id: I601c35a6bdab53c97e0742c77550e6bbe220d546
Signed-off-by: Namyoon Woo <namyoon@chromium.org>