Update patch set 1
Patch Set 1:

Andrey and I talked and since this will be fixed by the nvmem update, but that won't make the next release he was thinking we should probably land a bug fix for this release even if it is temporary.

I was thinking it would probably be OK to land this as a temporary fix that will be removed when the nvmem update lands.

Patch-set: 1
Reviewer: Gerrit User 1265562 <1265562@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>