Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > Patch Set 9:
> > 
> > > Patch Set 9:
> > > 
> > > > Patch Set 8:
> > > > 
> > > > (6 comments)
> > > > 
> > > > I am testing CTS on my side.
> > > > I wonder if the magnetometer/lsm6dsm can sustain 100Hz ODR. I tested overnight at 30Hz. 
> > > > Anyway, at 100Hz, the magnetometer may not have time to integrate properly.
> > > 
> > > Let me know how your CTS testing goes.  I tried my adjustment to the sample discard algorithm, but meep's tests are running about the same (I also ran it on phaser to ensure it didn't regress, and it doesn't.  So the change itself is at least okay for the non-magnetometer case. 
> > >  I can post it if you're interested).
> > > 
> > > I see test failures in the 50Hz and 25Hz range as well.  It looks like nearly every test failure measures the frequency around either 12.8 or 25 Hz.  
> > > Requested "CrosEC Compass" at 25.00hz (expecting between 22.50Hz and 55.00Hz, measured 12.80H
> > > z)
> > > Requested "GeoMag Rotation Vector Sensor" at 50.00hz (expecting between 45.00Hz and 457.57Hz, measured 25.61Hz)
> > > Requested "CrosEC Compass" at 50.00hz (expecting between 45.00Hz and 110.00Hz, measured 12.73Hz)
> > 
> > Same thing.
> > I am looking at 2 CTS issues:
> > - presenting magnetometer to Android.
> > [I don't like how the sensors are found currently: normally, the sensor presence should be identified using board information, not if the sensor are present or not: see https://chrome-internal-review.googlesource.com/c/chromeos/overlays/overlay-coral-private/+/540559/2/chromeos-base/chromeos-bsp-coral-private/files/common/arc%252B%252B/board_hardware_features#33]
> > - FIFO seems to send magnetometer data at 12.5Hz only. What ever non zero frequency we give.
> 

I believe this tracks with the application notes as I read them.  The FIFO trigger will be tied to the accelerometer and gyro ODRs with this equation:

F(odr) = min ( MAX(ODR_XL, ODR_G), ODR_FIFO )

So even if we've set ODR_FIFO to (say) 25Hz but the accelerometer is still only 13Hz, the FIFO trigger will operate at 13Hz.  This is in the "Setting the FIFO trigger, FIFO ODR and decimation factors" section of the application notes.

This seems to track with the Sensor Hub mode description.  There is a mention of an alternative configuration we could have done if we'd wired up the compass interrupt to the INT2 pin on the LSM6DSM as well:

"The sensor hub trigger signal can be synchronized with the accelerometer data-ready
signal (up to 104 Hz); alternatively, an external signal connected to the INT2 pin can be used as the sensor hub trigger"

Patch-set: 9
Reviewer: Gerrit User 1287197 <1287197@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>