Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> sorry it took me a while to get to this.
> 
> Having examine the code, it looks like the only place where pinvweaver_init() is called is the TPM_Startup callback. 

correct.

> 
> We know for a fact that there will be an nvmem_commit() invoked after TPM_Startup() completes, so there is no need in

we are fixing more than pinweaver_init behavior - we are fixing writevars (and other potential sources of nvmem_commit) behavior in general. if it happens while executing a command on the tpm task, it is possible that command also wrote something to nvmem and set g_updateNV already. in such case, g_updateNV shall be cleared on successful commit.

Rather than returning to fixing this issue in the future when we add something new that changes nvmem, we can address it now in a future-proof way.

Patch-set: 1
Reviewer: Gerrit User 1152319 <1152319@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>