LSM6: Adjust accel ODR to match magnetometer ODR needs
When magnetometer is behind the accelerometer and the accel is used as a
trigger to fetch the magnetometer readings and the magnetometer ODR is
greater than the accel ODR, adjust the accel ODR to match the
magnetometer ODR and then update the accel decimation settings so that
the final sample frequency match both the accel and magnetometer needs.

BUG=b:115587004
BRANCH=octopus
TEST=Boot to ChromeOS. Collect the magnetometer readings using ectool
motionsense command. Run CTS Sensor Test Suite

Change-Id: I92c63be484ca8835c57a3fd659822f73d7e7f43c
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>