Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> > Patch Set 1:
> > 
> > (1 comment)
> > 
> > sorry it took me a while to get to this.
> > 
> > Having examine the code, it looks like the only place where pinvweaver_init() is called is the TPM_Startup callback. 
> 
> correct.
> 
> > 
> > We know for a fact that there will be an nvmem_commit() invoked after TPM_Startup() completes, so there is no need in
> 
> we are fixing more than pinweaver_init behavior - we are fixing writevars (and other potential sources of nvmem_commit) behavior in general. if it happens while executing a command on the tpm task, it is possible that command also wrote something to nvmem and set g_updateNV already. in such case, g_updateNV shall be cleared on successful commit.
> 
> Rather than returning to fixing this issue in the future when we add something new that changes nvmem, we can address it now in a future-proof way.

well, in the future we're moving to an approach where writevars() is not even required, it would be easier to adjust to it by removing the call to copyvars() here and not touching everything else.

Again, just going and modifying an variable in a different library seems wrong to me. What if there is a change and some more modification to the cache happens after pinweaver_init() ran? It would be lost.

Patch-set: 1
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>