uut: drain EC console before programming
Cr50 doesn't handle the uart break conditions properly, so it will fill
up the EC uart with zeros when ECTX_CR50RX stays asserted. We assert
that signal to start uut programming, so the buffer is almost guaranteed
to be filled with garbage when uut programming starts. We were seeing
synchronization fail because the baudrate check would read a zero from
the buffer instead of the expected output. This change drains the EC
console before starting uut to get rid of the garbage that interferes
with programming.

stm32mon does the exact same thing. We should create common code for
initializing serial devices.

BUG=b:123775217
BRANCH=none
TEST=program cheza using ccd uut

Change-Id: I2df562a5b4513ada528e887e057e78a8f1f3054c
Signed-off-by: Mary Ruthven <mruthven@google.com>