hatch: Use CONFIG_TEMP_SENSOR_POWER_GPIO to prevent bad temp readings
This CL adds the config option CONFIG_TEMP_SENSOR_POWER_GPIO to make
sure that thermistor temperature readings are only taken when the
power rail for the thermistor circuit is enabled.

BRANCH=none
BUG=b:123900860
TEST=Verified that the the EC console message 'thermal SHUTDOWN' no
longer appears following an EC reboot as the ADC readings are now
ignored if the GPIO controlling the power rail is not enabled.

Change-Id: Ia6dced2f36ed6319c309868bf2d4db3493d29a01
Signed-off-by: Scott Collyer <scollyer@google.com>