Update patch set 1
Patch Set 1:

> Patch Set 1: -Verified -Commit-Queue
> 
> i think the watchdog boot loop was triggered by a malfunctioning
> (busted) servo v4.  with a "normal" servo v4, we still get a
> watchdog, but at least when we reset the TCPC on reboot,
> we do recover from the fault on the DuT side.
> 
> so, taking the watchdog seems like the lesser evil right now.

The boot loop can be triggered by unplugging DUT Power and leaving it unplugged, with the Servo still plugged into both the Host and DUT.

IIRC, any of the following is enough to bring the DUT out of that boot loop:
[A] Unplug Servo v4 CCD from DUT.
[B] Unplug Servo v4 CCD from Host.
[C] Reboot Servo v4 CCD.

While rebooting the Servo v4 CCD with all of Host, DUT Power, and DUT connected is enough to trigger the EC watchdog once, the DUT never gets stuck in a boot loop in that scenario.

All that said, I agree that letting the EC reboot (from watchdog) and therefore giving the best chance of recovering the TCPC is the least of evils here.

For now attempting to reset + reinitialize the TCPCs without an EC reset would be unnecessary complexity.  If Parade next week says they can't fix the TCPC in firmware, then we should reconsider whether it's worth having the EC reinitialize the TCPCs without the watchdog timeout.

Patch-set: 1