LSM6: Fix issue while setting Magnetometer ODR
While setting magnetometer ODR, accel ODR is restored through recursive
means. This causes the FIFO to be enabled as part of setting the accel
ODR while the magnetometer ODR is still to be configured. Fix the issue
by restoring the accel ODR in fifo_enable function instead of
recursively setting the accel ODR.

BUG=b:115587004
BRANCH=octopus
TEST=Boot to ChromeOS. Read the magnetometer samples using ectool.

Change-Id: I1c7b679b62c95277725c894242fa0e403d8cd37f
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>