driver: lsm6dsm: Spread FIFO datum properly
Previous code was not spreading properly.
Assuming:  Acc @ 52 Hz, Gyro @ 26 Hz Mag @ 13 Hz,
we should have:
 Gyro_0 | Acc_0 | Mag_0 | Acc_1 | Gyro_1 | Acc_2 | Acc_3

But driver was reading:
 Gyro_0 | Acc_0 | Mag_0 | Gyro_1 | Acc_1 | Acc_2 | Acc_3

It was working in 2 samples mode because will have:
 Gyro_0 | Acc_0 | Acc_1 | Acc_2 | ... | Acc_n
or
 Gyro_0 | Acc_0 | Gyro_1 |Gyro_2 | ... | Gyro_n

But the error uncovers once the magnetomer is added.
Using the ratio between sensors to decide if a given sensor is eligible
or another sensor is more suited to be selected.

BUG=b:124085261
BRANCH=none
TEST=Check samples produced by
android.hardware.cts.SensorIntegrationTests#testSensorsWithSeveralClients
belong to the right stream.

Change-Id: I2684637e11001cc488ba6c3336fe651195d64353
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>